+++ Solutions/Corvus.Json.CodeGeneration.202012/Corvus.Json.CodeGeneration.Draft202012/Generators/CodeGenerator.Array.Add.Partial.cs
@@ -739,7 +739,7 @@ public partial class CodeGeneratorArrayAdd
     {
         get
         {
-            return this.HasMaxItems ? this.TypeDeclaration.Schema().MaxItems : default;
+            return this.HasMaxItems ? (long)this.TypeDeclaration.Schema().MaxItems : default;
         }
     }
 
@@ -750,7 +750,7 @@ public partial class CodeGeneratorArrayAdd
     {
         get
         {
-            return this.HasMinItems ? this.TypeDeclaration.Schema().MinItems : default;
+            return this.HasMinItems ? (long)this.TypeDeclaration.Schema().MinItems : default;
         }
     }
 
@@ -1337,7 +1337,7 @@ public partial class CodeGeneratorArrayAdd
         {
             if (this.HasMaxContains)
             {
-                return this.TypeDeclaration.Schema().MaxContains;
+                return (long)this.TypeDeclaration.Schema().MaxContains;
             }
 
             return default;
@@ -1476,7 +1476,7 @@ public partial class CodeGeneratorArrayAdd
     {
         get
         {
-            return this.HasMinContains ? this.TypeDeclaration.Schema().MinContains : default;
+            return this.HasMinContains ? (long)this.TypeDeclaration.Schema().MinContains : default;
         }
     }
 
@@ -1520,7 +1520,7 @@ public partial class CodeGeneratorArrayAdd
     {
         get
         {
-            return this.TypeDeclaration.Schema().MaxProperties.IsNotUndefined() ? this.TypeDeclaration.Schema().MaxProperties : default;
+            return this.TypeDeclaration.Schema().MaxProperties.IsNotUndefined() ? (long)this.TypeDeclaration.Schema().MaxProperties : default;
         }
     }
 
@@ -1531,7 +1531,7 @@ public partial class CodeGeneratorArrayAdd
     {
         get
         {
-            return this.TypeDeclaration.Schema().MinProperties.IsNotUndefined() ? this.TypeDeclaration.Schema().MinProperties : default;
+            return this.TypeDeclaration.Schema().MinProperties.IsNotUndefined() ? (long)this.TypeDeclaration.Schema().MinProperties : default;
         }
     }
