+++ Solutions/Corvus.Json.CodeGeneration.6/Corvus.Json.CodeGeneration.Draft6/Generators/CodeGenerator.Validate.Format.Partial.cs
@@ -717,7 +717,7 @@ public partial class CodeGeneratorValidateFormat
     {
         get
         {
-            return this.HasMaxItems ? this.TypeDeclaration.Schema().MaxItems : default;
+            return this.HasMaxItems ? (long)this.TypeDeclaration.Schema().MaxItems : default;
         }
     }
 
@@ -728,7 +728,7 @@ public partial class CodeGeneratorValidateFormat
     {
         get
         {
-            return this.HasMinItems ? this.TypeDeclaration.Schema().MinItems : default;
+            return this.HasMinItems ? (long)this.TypeDeclaration.Schema().MinItems : default;
         }
     }
 
@@ -1358,7 +1358,7 @@ public partial class CodeGeneratorValidateFormat
     {
         get
         {
-            return this.TypeDeclaration.Schema().MaxProperties.IsNotUndefined() ? this.TypeDeclaration.Schema().MaxProperties : default;
+            return this.TypeDeclaration.Schema().MaxProperties.IsNotUndefined() ? (long)this.TypeDeclaration.Schema().MaxProperties : default;
         }
     }
 
@@ -1369,7 +1369,7 @@ public partial class CodeGeneratorValidateFormat
     {
         get
         {
-            return this.TypeDeclaration.Schema().MinProperties.IsNotUndefined() ? this.TypeDeclaration.Schema().MinProperties : default;
+            return this.TypeDeclaration.Schema().MinProperties.IsNotUndefined() ? (long)this.TypeDeclaration.Schema().MinProperties : default;
         }
     }
